requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
langchain-openai
langchain-chroma
langchain
pydantic
PyPDF2
PyMuPDF
Pillow
pytesseract
lxml
openai
nest-asyncio
langgraph
